requirements.txt
selenium==4.15.0
undetected-chromedriver==3.5.4
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0
oauth2client==4.1.3
boto3==1.29.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
python-dotenv==1.0.0